LayoutTests/platform/ios-wk2/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-flexbox-expected.txt

1
2
3
4
5
6
7
8
9
1
2
3
4
5
6
7
8
9
1
2
3
4
5
6
7
8
9

1
2
3
4
5
6
7
8
9

FAIL Flex assert_equals: offsetTop expected 28 but got 8
FAIL Inline flex assert_equals: offsetTop expected 124 but got 104
